test: move #12683 coverage into functional suite
Relocate constructor-preset relation hydration regression from
test/github-issues/12683 to test/functional/relations (preferred
location per project guidance) and query/assert using saved entity
ids instead of hardcoded primary keys.
