+++ test/functional/relations/constructor-preset-relation-hydration/entity/Author.ts
@@ -1,0 +1,12 @@
+import { Column } from "../../../../../src/decorator/columns/Column"
+import { Entity } from "../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+
+@Entity()
+export class Author {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+}
