--- test/github-issues/12683/entity/Author.ts
@@ -1,12 +1,0 @@
-import { Column } from "../../../../src/decorator/columns/Column"
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../src/decorator/columns/PrimaryGeneratedColumn"
-
-@Entity()
-export class Author {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    name: string
-}
